[refactor] 약물관리 디렉토리 구조 변경
