+++ components/Card/GroupMedicineCard.tsx
@@ -1,0 +1,16 @@
+import React from "react";
+import { StyleSheet, Text, View } from "react-native";
+
+interface GroupMedicineCardProps {}
+
+const GroupMedicineCard = () => {
+  return (
+    <View>
+      <Text>GroupMedicineCard</Text>
+    </View>
+  );
+};
+
+export default GroupMedicineCard;
+
+const styles = StyleSheet.create({});
